% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editData.R
\name{editDataServer}
\alias{editDataServer}
\title{Server for the editData module}
\usage{
editDataServer(id, data1 = NULL, data2 = NULL)
}
\description{
This function provides server for the data edit table.
}
