% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{Shiny server}
\usage{
server(input, output, session)
}
\arguments{
\item{input}{Inputs obtained from the user}

\item{output}{Computed outputs}

\item{session}{Shiny stuff}
}
\description{
This function contains server parts of shiny modules
}
